simplify cache invalidation
according to the AWS documentation (https://aws.amazon.com/cloudfront/pricing/ and https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/Invalidation.html) the cost for invalidation is based on the number of paths submitted for invalidation and not the total number of URLs invalidated. So it's easier (and cheaper) to invalidate everything every time instead of trying to be clever and only invalidating only the files that have changed.